requirements.txt
Adafruit-GPIO==1.0.3
influxdb==5.3.1
RPi.GPIO==0.7.1
pyserial==3.5